requirements.txt
fastapi
uvicorn
google-genai
websockets
python-dotenv
python-multipart
aiohttp
certifi
google-auth